backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Redis
redis==5.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket
python-socketio==5.11.0

# Background Tasks
apscheduler==3.10.4

# HTTP Client (for bots)
httpx==0.26.0

# Utilities
python-slugify==8.0.1
faker==22.0.0
Pillow==10.2.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
